scripts/receipt/requirements.txt
jinja2
qrcode[pil]
pillow
